refactor: 💡 Extract middlewareInstructionToHistoricInstruction
